+++ app_plus/src/main/java/org/gkisalatiga/plus/lib/Exceptions.kt
@@ -6,4 +6,5 @@
 
 package org.gkisalatiga.plus.lib
 
+class FileNotDownloadableException(errorMsg: String = "The requested target remote file seems to be corrupt or cannot be downloaded directly.") : Exception(errorMsg)
 class InvalidConnectionTestStringException(errorMsg: String = "Downloaded string does not match with the app's test string.") : Exception(errorMsg)